Refactor and reorganize project testing & utility scripts
- Reorganize scripts into logical directories:
  - scripts/devtools/: Developer utilities for test management
  - scripts/ci/: CI pipeline scripts for automated testing

- Refactor assemble_test.py → test_manager.py with subcommand CLI:
  - test_manager.py init: Initialize new test files from templates
  - test_manager.py assemble: Assemble single test with custom regions
  - test_manager.py assemble-all-sources: Batch assembly of all tests
  - test_manager.py clean: Remove test artifacts with confirmation

- Refactor generate_verilog_fixtures.py → build_all_fixtures.py:
  - Streamlined fixture generation for all testbenches
  - Uses test_manager.py for consistent assembly interface

- Refactor run_tests.py → run_test_suite.py:
  - Preserved all existing functionality
  - Updated paths for new script organization

- Update CI workflow (.github/workflows/ci.yml):
  - Updated script paths to use new organization
  - Maintained all existing functionality and error handling

- Update project documentation:
  - CLAUDE.md: Updated with new script commands and organization
  - scripts/README.md: Comprehensive guide for all scripts and workflows

- Remove obsolete files and directories:
  - Cleaned up old scripts/python/ directory
  - Removed outdated documentation files
